Add custom FlashHead models for transformers (#6)
Added custom models for FlashHead (Llama, Gemma and Qwen) so that
using the model does not accidentally use default generation
pipelines or vLLM inference.

Also updated the requirements for installation to remove flashinfer
so that the installation works for CPU and GPU.

Added minor improvements to the code quality and exposed the 
argument for GPU memory utilization in the demo.